Zoom changes to the background
Added scaling to the camera to achieve a more minimal effect on the change of the camera size displaying the backgrounds at different sizes. If anyone wants to fine tune it to numbers that would be exact, feel free, but I felt as if a slight change in background size is perfectly reasonable.