Node Backgrounds And Node Scaling
Scale the nodes to reflect the change in local scale by the network parent, as well as place a background node as a child of  the node so that it can be displayed and seen over the background. Will be changed to the white sprites and have the backgrounds removed in a future change.